requirements.txt
attrs==23.2.0
cmd2==2.4.3
colorama==0.4.6
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
PrettyPrintTree==2.0.0
pyperclip==1.8.2
pytest==8.1.1
wcwidth==0.2.13
